+++ rest_framework/routers.py
@@ -16,6 +16,7 @@ For example, you might have a `urls.py` that looks something like this:
 from __future__ import unicode_literals
 
 import itertools
+import warnings
 from collections import OrderedDict, namedtuple
 
 from django.conf.urls import url
@@ -30,9 +31,30 @@ from rest_framework.schemas.views import SchemaView
 from rest_framework.settings import api_settings
 from rest_framework.urlpatterns import format_suffix_patterns
 
-Route = namedtuple('Route', ['url', 'mapping', 'name', 'initkwargs'])
-DynamicDetailRoute = namedtuple('DynamicDetailRoute', ['url', 'name', 'initkwargs'])
-DynamicListRoute = namedtuple('DynamicListRoute', ['url', 'name', 'initkwargs'])
+Route = namedtuple('Route', ['url', 'mapping', 'name', 'detail', 'initkwargs'])
+DynamicRoute = namedtuple('DynamicRoute', ['url', 'name', 'detail', 'initkwargs'])
+
+
+class DynamicDetailRoute(object):
+    def __new__(cls, url, name, initkwargs):
+        warnings.warn(
+            "`DynamicDetailRoute` is pending deprecation and will be removed in 3.10 "
+            "in favor of `DynamicRoute`, which accepts a `detail` boolean. Use "
+            "`DynamicRoute(url, name, True, initkwargs)` instead.",
+            PendingDeprecationWarning, stacklevel=2
+        )
+        return DynamicRoute(url, name, True, initkwargs)
+
+
+class DynamicListRoute(object):
+    def __new__(cls, url, name, initkwargs):
+        warnings.warn(
+            "`DynamicListRoute` is pending deprecation and will be removed in 3.10 in "
+            "favor of `DynamicRoute`, which accepts a `detail` boolean. Use "
+            "`DynamicRoute(url, name, False, initkwargs)` instead.",
+            PendingDeprecationWarning, stacklevel=2
+        )
+        return DynamicRoute(url, name, False, initkwargs)
 
 
 def escape_curly_brackets(url_path):
@@ -42,18 +64,6 @@ def escape_curly_brackets(url_path):
     if ('{' and '}') in url_path:
         url_path = url_path.replace('{', '{{').replace('}', '}}')
     return url_path
-
-
-def replace_methodname(format_string, methodname):
-    """
-    Partially format a format_string, swapping out any
-    '{methodname}' or '{methodnamehyphen}' components.
-    """
-    methodnamehyphen = methodname.replace('_', '-')
-    ret = format_string
-    ret = ret.replace('{methodname}', methodname)
-    ret = ret.replace('{methodnamehyphen}', methodnamehyphen)
-    return ret
 
 
 def flatten(list_of_lists):
@@ -103,14 +113,15 @@ class SimpleRouter(BaseRouter):
                 'post': 'create'
             },
             name='{basename}-list',
+            detail=False,
             initkwargs={'suffix': 'List'}
         ),
-        # Dynamically generated list routes.
-        # Generated using @list_route decorator
-        # on methods of the viewset.
-        DynamicListRoute(
-            url=r'^{prefix}/{methodname}{trailing_slash}$',
-            name='{basename}-{methodnamehyphen}',
+        # Dynamically generated list routes. Generated using
+        # @action(detail=False) decorator on methods of the viewset.
+        DynamicRoute(
+            url=r'^{prefix}/{url_path}{trailing_slash}$',
+            name='{basename}-{url_name}',
+            detail=False,
             initkwargs={}
         ),
         # Detail route.
@@ -123,13 +134,15 @@ class SimpleRouter(BaseRouter):
                 'delete': 'destroy'
             },
             name='{basename}-detail',
+            detail=True,
             initkwargs={'suffix': 'Instance'}
         ),
-        # Dynamically generated detail routes.
-        # Generated using @detail_route decorator on methods of the viewset.
-        DynamicDetailRoute(
-            url=r'^{prefix}/{lookup}/{methodname}{trailing_slash}$',
-            name='{basename}-{methodnamehyphen}',
+        # Dynamically generated detail routes. Generated using
+        # @action(detail=True) decorator on methods of the viewset.
+        DynamicRoute(
+            url=r'^{prefix}/{lookup}/{url_path}{trailing_slash}$',
+            name='{basename}-{url_name}',
+            detail=True,
             initkwargs={}
         ),
     ]
@@ -160,57 +173,47 @@ class SimpleRouter(BaseRouter):
         # converting to list as iterables are good for one pass, known host needs to be checked again and again for
         # different functions.
         known_actions = list(flatten([route.mapping.values() for route in self.routes if isinstance(route, Route)]))
+        extra_actions = viewset.get_extra_actions()
 
-        # Determine any `@detail_route` or `@list_route` decorated methods on the viewset
-        detail_routes = []
-        list_routes = []
-        for methodname in dir(viewset):
-            attr = getattr(viewset, methodname)
-            httpmethods = getattr(attr, 'bind_to_methods', None)
-            detail = getattr(attr, 'detail', True)
-            if httpmethods:
-                # checking method names against the known actions list
-                if methodname in known_actions:
-                    raise ImproperlyConfigured('Cannot use @detail_route or @list_route '
-                                               'decorators on method "%s" '
-                                               'as it is an existing route' % methodname)
-                httpmethods = [method.lower() for method in httpmethods]
-                if detail:
-                    detail_routes.append((httpmethods, methodname))
-                else:
-                    list_routes.append((httpmethods, methodname))
+        # checking action names against the known actions list
+        not_allowed = [
+            action.__name__ for action in extra_actions
+            if action.__name__ in known_actions
+        ]
+        if not_allowed:
+            msg = ('Cannot use the @action decorator on the following '
+                   'methods, as they are existing routes: %s')
+            raise ImproperlyConfigured(msg % ', '.join(not_allowed))
 
-        def _get_dynamic_routes(route, dynamic_routes):
-            ret = []
-            for httpmethods, methodname in dynamic_routes:
-                method_kwargs = getattr(viewset, methodname).kwargs
-                initkwargs = route.initkwargs.copy()
-                initkwargs.update(method_kwargs)
-                url_path = initkwargs.pop("url_path", None) or methodname
-                url_path = escape_curly_brackets(url_path)
-                url_name = initkwargs.pop("url_name", None) or url_path
-                ret.append(Route(
-                    url=replace_methodname(route.url, url_path),
-                    mapping={httpmethod: methodname for httpmethod in httpmethods},
-                    name=replace_methodname(route.name, url_name),
-                    initkwargs=initkwargs,
-                ))
+        # partition detail and list actions
+        detail_actions = [action for action in extra_actions if action.detail]
+        list_actions = [action for action in extra_actions if not action.detail]
 
-            return ret
-
-        ret = []
+        routes = []
         for route in self.routes:
-            if isinstance(route, DynamicDetailRoute):
-                # Dynamic detail routes (@detail_route decorator)
-                ret += _get_dynamic_routes(route, detail_routes)
-            elif isinstance(route, DynamicListRoute):
-                # Dynamic list routes (@list_route decorator)
-                ret += _get_dynamic_routes(route, list_routes)
+            if isinstance(route, DynamicRoute) and route.detail:
+                routes += [self._get_dynamic_route(route, action) for action in detail_actions]
+            elif isinstance(route, DynamicRoute) and not route.detail:
+                routes += [self._get_dynamic_route(route, action) for action in list_actions]
             else:
-                # Standard route
-                ret.append(route)
+                routes.append(route)
 
-        return ret
+        return routes
+
+    def _get_dynamic_route(self, route, action):
+        initkwargs = route.initkwargs.copy()
+        initkwargs.update(action.kwargs)
+
+        url_path = escape_curly_brackets(action.url_path)
+
+        return Route(
+            url=route.url.replace('{url_path}', url_path),
+            mapping={http_method: action.__name__
+                     for http_method in action.bind_to_methods},
+            name=route.name.replace('{url_name}', action.url_name),
+            detail=route.detail,
+            initkwargs=initkwargs,
+        )
 
     def get_method_map(self, viewset, method_map):
         """
@@ -281,6 +284,7 @@ class SimpleRouter(BaseRouter):
                 initkwargs = route.initkwargs.copy()
                 initkwargs.update({
                     'basename': basename,
+                    'detail': route.detail,
                 })
 
                 view = viewset.as_view(mapping, **initkwargs)