+++ rest_framework/viewsets.py
@@ -19,12 +19,17 @@ automatically.
 from __future__ import unicode_literals
 
 from functools import update_wrapper
+from inspect import getmembers
 
 from django.utils.decorators import classonlymethod
 from django.views.decorators.csrf import csrf_exempt
 
 from rest_framework import generics, mixins, views
 from rest_framework.reverse import reverse
+
+
+def _is_extra_action(attr):
+    return hasattr(attr, 'bind_to_methods')
 
 
 class ViewSetMixin(object):
@@ -50,6 +55,9 @@ class ViewSetMixin(object):
         # The suffix initkwarg is reserved for displaying the viewset type.
         # eg. 'List' or 'Instance'.
         cls.suffix = None
+
+        # The detail initkwarg is reserved for introspecting the viewset type.
+        cls.detail = None
 
         # Setting a basename allows a view to reverse its action urls. This
         # value is provided by the router through the initkwargs.
@@ -112,8 +120,7 @@ class ViewSetMixin(object):
 
     def initialize_request(self, request, *args, **kwargs):
         """
-        Set the `.action` attribute on the view,
-        depending on the request method.
+        Set the `.action` attribute on the view, depending on the request method.
         """
         request = super(ViewSetMixin, self).initialize_request(request, *args, **kwargs)
         method = request.method.lower()
@@ -134,6 +141,13 @@ class ViewSetMixin(object):
         kwargs.setdefault('request', self.request)
 
         return reverse(url_name, *args, **kwargs)
+
+    @classmethod
+    def get_extra_actions(cls):
+        """
+        Get the methods that are marked as an extra ViewSet `@action`.
+        """
+        return [method for _, method in getmembers(cls, _is_extra_action)]
 
 
 class ViewSet(ViewSetMixin, views.APIView):